+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Digital Shopfront CMS Security Changelog
 
+## 1.0.2 - April 21, 2025
+
+- Fixed an issue with running the `kses()` function.
+- Added in tests and a GitLab pipeline.
+
 ## 1.0.1 - April 20, 2025
 
 - Removed unnecessary files from the published package.